+++ interface/src/components/note-block/note-block.tsx
@@ -1,11 +1,21 @@
-import { useDragContext } from '../../contexts/drag-context';
+import { useState } from 'react';
+import { useDragContext } from '../../contexts/use-drag-context';
 import './note-block.css';
+import { NoteEvent } from 'midi-writer-js/build/types/midi-events/note-event';
 
 function NoteBlock() {
     const { isDragging, setIsDragging } = useDragContext();
+    const [note,] = useState<NoteEvent | null>(null);
+
+    const handleDragStart = (event: React.DragEvent) => {
+        setIsDragging!(true);
+        // You can set the note here if needed
+        // setNote({ ... });
+        event.dataTransfer.setData('text/plain', JSON.stringify(note));
+    };
 
     return (
-        <div draggable={!isDragging} onDragStart={() => setIsDragging(true)} onDragEnd={() => setIsDragging(false)} className='note-block'>
+        <div draggable={!isDragging} onDragStart={handleDragStart} className='note-block'>
             <p>test</p>
         </div>
     )
