+++ interface/src/components/staff/staff.tsx
@@ -1,4 +1,4 @@
-import { useDragContext } from "../../contexts/drag-context";
+import { useDragContext } from "../../contexts/use-drag-context";
 import NoteBlock from "../note-block/note-block";
 import './staff.css';
