+++ interface/src/contexts/drag-context-provider.tsx
@@ -1,17 +1,15 @@
-import { useState, type ReactNode, } from "react";
+import type { NoteEvent } from "midi-writer-js/build/types/midi-events/note-event";
+import { useState, type ReactNode } from "react";
 import { DragContext } from "./drag-context";
 
-interface DragProviderProps {
-    children: ReactNode;
-}
-
-export const DragProvider = ({ children }: DragProviderProps) => {
+export const DragProvider = ({ children }: { children: ReactNode }) => {
     const [isDragging, setIsDragging] = useState(false);
+    const [note, setNote] = useState<NoteEvent | null>(null);
 
     return (
-        <DragContext.Provider value={{ isDragging, setIsDragging }}>
+        <DragContext.Provider value={{ isDragging, note, setIsDragging, setNote }
+        }>
             {children}
         </DragContext.Provider>
     );
 };
-
